requirements.txt
aiohttp
aiosignal
annotated-types
antlr4-python3-runtime
async-timeout
attrs
backoff
beautifulsoup4
cachetools
certifi
cffi
chardet
charset-normalizer
click
coloredlogs
contourpy
cryptography
cycler
dataclasses-json
dataclasses-json-speakeasy
Deprecated
docstring_parser
effdet
emoji
filelock
filetype
flatbuffers
fonttools
frozenlist
fsspec
google-api-core
google-auth
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-core
google-cloud-resource-manager
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
greenlet
grpc-google-iam-v1
grpcio
grpcio-status
huggingface-hub
humanfriendly
idna
iopath
Jinja2
joblib
jsonpatch
jsonpath-python
jsonpointer
kiwisolver
langchain
langchain-community
langchain-core
langchain-google-vertexai
langchain-text-splitters
langchainhub
langdetect
langsmith
layoutparser
lxml
MarkupSafe
marshmallow
matplotlib
mpmath
multidict
mypy-extensions
networkx
nltk
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
omegaconf
onnx
onnxruntime
opencv-python
orjson
packaging
pandas
pdf2image
pdfminer.six==20231228
pdfplumber
pikepdf
pillow
pillow_heif
pinecone-client
portalocker
proto-plus
protobuf
pyasn1
pyasn1_modules
pycocotools
pycparser
pycryptodome
pydantic
pydantic_core
pyparsing
pypdf
pypdfium2
pytesseract
python-dateutil
python-dotenv
python-iso639
python-magic
python-multipart
pytz
PyYAML
rapidfuzz
regex
requests
rsa
safetensors
scipy
shapely
six
soupsieve
SQLAlchemy
sympy
tabulate
tenacity
timm
tokenizers
torch
torchvision
tqdm
transformers
triton
types-protobuf
types-requests
typing-inspect
typing_extensions
tzdata
unstructured
unstructured-client
unstructured-inference
unstructured.pytesseract
urllib3
wrapt
yarl
